requirements.txt
aiohttp>=3.8.0
numpy>=1.21.0
python-dotenv>=0.19.0
pytest>=7.0.0
pytest-asyncio>=0.20.0
pytest-cov>=4.0.0
rouge-score>=0.1.2